Extract useBackgroundColorTween into a shared module
PayoutTable.tsx defined the OKLab color tween hook privately; pulling it
into src/app/components/ui/useBackgroundColorTween.ts lets ArenaTableBody
and DropDownTable reuse it instead of duplicating the same logic.
